backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
sqlalchemy==2.0.36
greenlet==3.1.1
aiosqlite==0.20.0
scrapli[asyncssh]==2024.7.30
scrapli-community==2024.7.30
ntc-templates==9.0.0
textfsm==1.1.3
python-dotenv==1.0.1
httpx==0.28.1
PyYAML==6.0.2
